refactor: 파일 구조 변경 및 분리
