+++ src/LottoGame.js
@@ -1,0 +1,10 @@
+// makeLottoNumbers() {
+//   const lottoNumbers = [];
+
+//   while (lottoNumbers.length < LOTTO.numbersLength) {
+//     const number = generateRandomNumber(LOTTO.minNumber, LOTTO.maxNumber);
+//     if (!lottoNumbers.includes(number)) lottoNumbers.push(number);
+//   }
+
+//   return lottoNumbers.sort((a, b) => a - b);
+// },
