+++ src/domain/models/Lotto.js
